Refactor/aprimoramento das classes controllers e services

+++ src/main/java/br/com/listaDeComprasApi/domain/service/ProdutoService.java
@@ -4,80 +4,95 @@ import br.com.listaDeComprasApi.domain.exception.EntidadeNaoEncontradaException;
 import br.com.listaDeComprasApi.domain.exception.NegocioException;
 import br.com.listaDeComprasApi.domain.model.ListaDeCompras;
 import br.com.listaDeComprasApi.domain.model.Produto;
-import br.com.listaDeComprasApi.domain.repository.ListaDeComprasRepository;
 import br.com.listaDeComprasApi.domain.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Objects;
 
 @Service
 public class ProdutoService {
     @Autowired
     ProdutoRepository produtoRepository;
-
     @Autowired
     ListaDeComprasService listaDeComprasService;
-
     private static final String MSG_PRODUTO_NAO_ENCOTRADO = "Não existe registro de produto com o id informado.";
-
     private static final String MSG_PRODUTO_COM_MESMO_NOME = "Já existe registro de produto com o nome inserido.";
+    private static final String MSG_LISTA_NAO_ENCOTRADA = "Não existe registro de lista com o id informado.";
 
     @Transactional
     public Produto buscarPorId (Long idProduto) {
         return produtoRepository.findById(idProduto).orElseThrow(() -> new EntidadeNaoEncontradaException(MSG_PRODUTO_NAO_ENCOTRADO));
-
     }
 
     @Transactional
     public Page<Produto> buscarProdutos (Pageable pageable, Long idListaDeCompras) {
         return produtoRepository.findAllProdutosByIdListaDeCompras(pageable, idListaDeCompras);
-
     }
 
     @Transactional
     public Produto salvar (Produto produto){
-        ListaDeCompras listaDeCompras = listaDeComprasService.buscarPorId(produto.getListaDeCompras().getIdListaDeCompras());
-
-        produto.setListaDeCompras(listaDeCompras);
-
+        buscaEAssociaListaDeComprasAProduto(produto);
         verificarSeExisteProdutoComMesmoNome(produto);
 
         return produtoRepository.save(produto);
     }
 
-    private void verificarSeExisteProdutoComMesmoNome(Produto produto) {
-        Produto resultadoBusca;
+    private void buscaEAssociaListaDeComprasAProduto (Produto produto) {
+        ListaDeCompras listaDeCompras = listaDeComprasService.buscarPorId(produto.getListaDeCompras().getIdListaDeCompras());
 
-        if (!Objects.isNull(produto.getIdProduto())){
-            resultadoBusca = produtoRepository.findByNameAndIdProdutoAndIdLista(produto.getNome(), produto.getListaDeCompras().getIdListaDeCompras(), produto.getIdProduto());
+        verificarSeListaRetornadaNaoEVazia(listaDeCompras);
 
-        } else {
-            resultadoBusca = produtoRepository.findByNameAndIdLista(produto.getNome(), produto.getListaDeCompras().getIdListaDeCompras());
+        produto.setListaDeCompras(listaDeCompras);
+    }
 
-        }
-
-        if(!Objects.isNull(resultadoBusca)) {
-            throw new NegocioException(MSG_PRODUTO_COM_MESMO_NOME);
-
+    private void verificarSeListaRetornadaNaoEVazia(ListaDeCompras listaDeCompras) {
+        if (isListaNula(listaDeCompras)) {
+            throw new EntidadeNaoEncontradaException(MSG_LISTA_NAO_ENCOTRADA);
         }
     }
 
-    @Transactional
-    public void remover(Long idProduto) {
-        try {
-            produtoRepository.deleteById(idProduto);
-            produtoRepository.flush();
+    private boolean isListaNula(ListaDeCompras listaDeCompras){
+        return Objects.isNull(listaDeCompras);
+    }
 
-        } catch (EmptyResultDataAccessException e) {
-            throw new EntidadeNaoEncontradaException(MSG_PRODUTO_NAO_ENCOTRADO);
+    private void verificarSeExisteProdutoComMesmoNome(Produto produto) {
+        Boolean existeProdutoComMesmoNome;
 
+        if (isIdNaoNulo(produto)){
+            existeProdutoComMesmoNome = produtoRepository.findByNameAndIdProdutoAndIdLista(produto.getNome(), produto.getListaDeCompras().getIdListaDeCompras(), produto.getIdProduto());
+
+        } else {
+            existeProdutoComMesmoNome = produtoRepository.findByNameAndIdLista(produto.getNome(), produto.getListaDeCompras().getIdListaDeCompras());
         }
+
+        if(existeProdutoComMesmoNome) {
+            throw new NegocioException(MSG_PRODUTO_COM_MESMO_NOME);
+        }
+    }
+
+    private boolean isIdNaoNulo(Produto produto){
+        return !Objects.isNull(produto.getIdProduto());
+    }
+
+    @Transactional
+    public ResponseEntity <Void> remover(Long idProduto) {
+        if (isProdutoInexistente(idProduto)) {
+            throw new EntidadeNaoEncontradaException(MSG_PRODUTO_NAO_ENCOTRADO);
+        }
+
+        produtoRepository.deleteById(idProduto);
+        produtoRepository.flush();
+
+        return ResponseEntity.noContent().build();
+    }
+
+    private boolean isProdutoInexistente(Long idProduto) {
+        return !produtoRepository.existsById(idProduto);
     }
 
 }
